+++ corpan/plugins/tauri-plugin-iap/src/mobile.rs
@@ -105,4 +105,11 @@ impl<R: Runtime> Iap<R> {
             .await
             .map_err(Into::into)
     }
+
+    pub async fn request_review(&self) -> crate::Result<()> {
+        self.0
+            .run_mobile_plugin_async("requestReview", ())
+            .await
+            .map_err(Into::into)
+    }
 }